# perf(weave): add purpose built endpoint for project data