## Description

<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->

[WB-24603](https://wandb.atlassian.net/browse/WB-24603)

Adds a purpose built endpoint for checking if a project has traces. Query `project_check` with a request payload of the `project_id` and get back a `has_data` bool. 

Fast follow in core: https://github.com/wandb/core/pull/28989

## Testing

- Unit test
- Manual performance testing, in high trace environment:

| Environment | Elapsed Time | Rows Read  | Data Read | Peak Memory   | Error |
|-------------|--------------|------------|-----------|----------------|--------|
| Prod        | 3.5s         | 53,890,051 | –         | 31.94 GiB      | True   |
| Branch      | 0.008s       | 15,631     | 1.24 MB   | 126.63 MiB     | False  |

In prod, this query is run 1.2 million times a week, so even modest improvements in speed might actually have an impact on clickhouse performance. 

| Env | Runs | Errors | Avg. Read Rows | Avg. Memory Usage | p5 | p90 | p99 |
|-------------|------------------|--------|----------------|-------------------|-------------|-------------|-------------|
| Prod        | 1,169,821        | 0      | 30,333.25      | 95.86 MiB         | 0.019s      | 0.04s       | 0.541s      |

[WB-24603]: https://wandb.atlassian.net/browse/WB-24603?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ