@gtarpenning 2025-04-28
> Can you explain the use-case for this endpoint?

@neutralino1 

The frontend pr should clarify: https://github.com/wandb/weave/pull/4269

We use it for determining whether or not to show the weave UI when rendering the core models UI.