@neutralino1 2025-04-28
I see. Is there any chance we could roll this into the `/calls/stats` endpoint? Like maybe with a special query arg...