@gtarpenning 2025-04-28
> I see. Is there any chance we could roll this into the `/calls/stats` endpoint? Like maybe with a special query arg...

1. yes, reasonable request. I initially wanted to do this with like a `LIMIT` param. You could hit the stats endpoint with a limit, which would be useful in more than just this case. For example we could show in the UI `100+` for large expensive queries instead of returning `102,301` or whatever. 
2. However, fundamentally this query is different. It checks both calls and **objects**, and uses a very custom query construction. I think there are arguments to be made that it should be separate. Because of the objects bit it doesn't exactly fit under `calls/stats`
3. I think we will eventually want a query that returns the lastest runs in the `x` last projects. This means that we will probably eventually want a `project/` endpoint anyway. It might be `project/latest` etc. This query right now is still too expensive to work on every project though, otherwise I would have gone forward with that.

I can write up a doc if we think its an invasive enough change. In general though we are only exposing this at the service layer, so it won't be easily accessible to users, so if/when we change this endpoint it won't require back-compat etc.