@neutralino1 2025-04-28
Re: calls+object – what if we modified both `/calls/query_stats` and `/objects/query_stats` and queried both in parallel?

Although I just realized we don't have `/objects/query_stats`. 🙃