@gtarpenning 2025-04-28
> Re: calls+object – what if we modified both `/calls/query_stats` and `/objects/query_stats` and queried both in parallel?
> 
> Although I just realized we don't have `/objects/query_stats`. 🙃

@neutralino1 

we could do this, totally an option. One additional wrinkle though is that for `calls/query_stats` I will still be implementing custom handling for `LIMIT 1`, or adding a special flag for this query. The existing calls stats query does not work for mega large projects even with `LIMIT 1`. 

So, the options are: 

Path 1 (this pr):
- custom endpoint for just this very specific case

Path 2:
- add LIMIT to `calls/stats`, which will be useful in other cases
- add special handling for `LIMIT 1` or special flag for this specific case
- add `objects/stats` endpoint with `LIMIT` handling and maybe also special case for `LIMIT 1`

Honestly I could go either way, both seem annoying.