@neutralino1 2025-04-28
I understand Path 2 is a little more work. My goal here is to limit custom endpoints and instead have powerful and flexible endpoints operating on primitives (calls, objects, etc.).
We will have to enhance the stats endpoint anyway for the purpose of trace plots and dashboarding. So maybe this work would align with that.