@gtarpenning 2025-04-28
closing in favor of https://github.com/wandb/weave/pull/4331